Use memcpy, etc. from libc crate n xmlparse.rs

+++ src/expat_config_h.rs
@@ -1,2 +1,2 @@
-use libc::c_int;
-pub const XML_CONTEXT_BYTES: c_int = 1024;
+
+pub const XML_CONTEXT_BYTES: isize = 1024;
